refactor(core-wallet-ui-components): extract common code from wallet form
